# adding guards for LLVM optimization